Update skip_consent.md
If you reauthenticate an active session using max_age 0, the _session cookie id changes and the old session row is deleted
in the database but somehow the existing grant id gets copied onto the new session. The grant's expiry isn't updated, however.
You end up with a session length longer than the grant's length. In the case of skipped consent, this ends up with the interaction policy eventually landing you on the interaction start handler with prompt 'consent' (once the grant has expired but your session is still valid). that won't be implemented (because you are trying to skip it) so things error.